fix exchange rates y permito decimal

+++ WalletKata/Services/Interfaces/IWalletService.cs
@@ -2,9 +2,9 @@
 {
     public interface IWalletService
     {
-        Task Deposit(int userId, string currencyCode, int amount);
-        Task Withdraw(int userId, string currencyCode, int amount);
-        Task<Dictionary<string, int>> GetBalance(long userId);
-        Task Exchange(int userId, string sourceCurrencyCode, string targetCurrencyCode, int amount);
+        Task Deposit(int userId, string currencyCode, decimal amount);
+        Task Withdraw(int userId, string currencyCode, decimal amount);
+        Task<Dictionary<string, decimal>> GetBalance(long userId);
+        Task Exchange(int userId, string sourceCurrencyCode, string targetCurrencyCode, decimal amount);
     }
 }
